Criação da base de dados, leia.txt
