--- Nova pasta - Copia/conexao.php
@@ -1,14 +1,0 @@
-<?php
-$servidor = "localhost";
-$usuario = "root";
-$senha = "";
-$dbname = "formulario-"
-
-
-
-// criar a conexao 
-
-
-$conn = mysqli_connect($servidor,$usuario, $senha, $dbname);
-
-?>
