+++ projetoNovo/Assets/CSS/sytle.css
@@ -1,0 +1,81 @@
+*{
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+    font-family: "Poppins", sans-serif;
+}
+
+body{
+    background-color: aquamarine;
+    width: 100%;
+    min-height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.text-one{
+
+    background-color: #fafafa;
+    border-radius: 13px;
+    margin: 14px; 
+    max-width: 600px;
+    width: 100%;
+    box-sizing:  0 3px 5px rgba(0,0,0,0.5);
+    overflow: hidden;
+
+}
+
+.header{
+    background:  linear-gradient(190deg, #3acbf0
+     0%, #8bb0ff 100% );
+     padding: 24px;
+     text-align:center;
+     color: #fff
+}
+
+.form{
+    padding: 18px;
+}
+
+.form-content{
+    margin-bottom: 8px;
+    padding-bottom: 18px;
+    position: relative;
+}
+.form-content input{
+    display: block;
+    width: 100%;
+    border-radius: 8px;
+    padding: 8px;
+    border: 2px solid #dfdfdf;
+}
+.form-content a {
+    position: absolute;
+    bottom: -8px;
+    left: 0;
+    visibility: hidden;
+}
+
+
+.form  button{
+    background-color: #00c3ff;
+    color: #fff;
+    width: 100%;
+    border-radius: 14px;
+    padding: 10px;
+    border: 0;
+    font-size: 16px;
+    cursor: pointer;
+    margin-top:14px
+}
+
+
+.form-content.error input{
+    border-color: red;
+
+}
+.form-content.error a{
+    color: red;
+    visibility: visible;
+}
