+++ projetoNovo/Model/usuario.php
@@ -1,0 +1,22 @@
+<?php
+    class Usuario{
+        private $matricula;
+        private $email;
+        private $ctt;
+        private $id;
+        private $nick;
+        private $nome;
+        private $email_corp;
+        private $ctt_corp;
+        private $senha;
+        private $rank;
+        private $dataCad;
+        private $Matricula;
+            public function solicitarCadastro(){
+
+            }
+            public function novoCadastro(){
+                
+            }
+    }
+?>
