+++ projetoNovo/View/index.html
@@ -1,0 +1,85 @@
+<!DOCTYPE html>
+<html lang="PT-br">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Formulário - Central</title>
+    <link rel="stylesheet" href="../Assets/CSS/sytle.css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="stylesheet" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Poppins:wgh@400;500;700&display=swap" rel="stylesheet">
+    <script src="../Assets/JS/script.js"></script>
+</head>
+<body>
+
+   
+
+
+    <div class="text-one">
+        <section class="header">
+            <h2>Formulário Central</h2>
+        </section>
+
+        <form id="form" class="form">
+            <div class="form-content">
+                <label for="username">Matrícula:  </label>
+                <input type="number"
+                id="username"
+                placeholder="Digite sua matrícula...  "/>
+
+                <a>Aqui vai uma mensagem de erro...</a>
+            </div>
+
+
+            <div class="form-content">
+                <label for="email">Email: </label>
+                <input type="email"
+                id="email"
+                placeholder="Digite seu email...  "/>
+
+                <a>Aqui vai uma mensagem de erro...</a>
+            </div>
+
+
+
+            <div class="form-content">
+                <label for="password">Senha:  </label>
+                <input type="password"
+                id="password"
+                placeholder="Digite sua senha...  "/>
+
+                <a>Aqui vai uma mensagem de erro...</a>
+            </div>
+
+            <div class="form-content  ">
+                <label for="password-confirmation">Confirmação de senha: </label>
+                <input 
+                type="password"
+                id="password-confirmation"
+                placeholder="Digite sua senha novamente...  "
+                />
+                <a>Aqui vai uma mensagem de erro...</a>
+
+
+            </div>
+
+            <textarea cols="40" rows="5">Área de texto!</textarea>
+
+            <div class="form-content">
+                <input type="file" />
+
+                <a>Aqui vai uma mensagem de erro...</a>
+            </div>
+
+            
+
+            <button type="submit"> Enviar</button>
+
+        </form>
+    </div>
+
+    <script src="./script.js"></script>
+    
+</body>
+</html>
