+++ projetoNovo/connection.php
@@ -1,0 +1,14 @@
+<?php
+$servidor = "localhost";
+$usuario = "root";
+$senha = "";
+$dbname = "formulario-";
+
+
+
+// criar a conexao 
+
+
+$conn = mysqli_connect($servidor,$usuario, $senha, $dbname);
+
+?>
